refactor: rename `approximators` -> `eval`
