+++ src/lib.rs
@@ -1,9 +1,6 @@
 extern crate itertools;
 extern crate ndarray;
 extern crate rand;
-
-pub extern crate spaces as geometry;
-
 extern crate serde;
 #[macro_use]
 extern crate serde_derive;
@@ -11,11 +8,13 @@ extern crate serde_derive;
 #[cfg(tests)]
 extern crate quickcheck;
 
+pub extern crate spaces as geometry;
+
 mod macros;
 mod utils;
 
 import_all!(lfa);
 
-pub mod approximators;
-pub mod basis;
 pub mod core;
+pub mod eval;
+pub mod basis;
